Move provider, reducer into their own files/folders
