+++ apps/react/demo/src/ui/Example/Context.tsx
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import type { ProviderValue } from "./types.js";
+import type { ProviderValue } from "./Provider/index.js";
 
 const Context = createContext<ProviderValue | undefined>(undefined);
