+++ apps/react/demo/src/ui/Example/Provider/index.ts
@@ -1,0 +1,2 @@
+export { default as Provider } from "./Provider.js";
+export * from "./types.js";
