+++ apps/react/demo/src/ui/Example/Reducer/index.ts
@@ -1,0 +1,2 @@
+export { default as Reducer } from "./Reducer.js";
+export * from "./types.js";
