+++ apps/react/demo/src/ui/Example/index.ts
@@ -1,10 +1,10 @@
-import type { ExampleControlsComponent } from "./types.js";
+import type { ExampleComponent } from "./types.js";
 
 export * from "./types.js";
 
-import Provider from "./Provider.js";
+import { Provider } from "./Provider/index.js";
 import { Controls } from "./common/Controls/index.js";
 import { Renderer } from "./common/Renderer/index.js";
-export const Example = Provider as ExampleControlsComponent;
+export const Example = Provider as ExampleComponent;
 Example.Controls = Controls;
 Example.Renderer = Renderer;
